Updated core to 4.1.2+ (Build: 20230322) + plugins to 2023-03-23

+++ local/o365/classes/observers.php
@@ -368,32 +368,29 @@ class observers {
             $userobject = $DB->get_record('local_o365_objects', ['type' => 'user', 'moodleid' => $userid]);
 
             if (empty($userobject)) {
-                // Skip field mapping if the user uses auth_oidc, and matching record is stored in local_o365_objects table.
-                if ($existinguserdata->auth != 'oidc') {
-                    // Create o365_object record if it does not exist.
-                    $tenant = utils::get_tenant_for_user($userid);
-                    $metadata = '';
-                    if (!empty($tenant)) {
-                        // Additional tenant - get ODB url.
-                        $odburl = utils::get_odburl_for_user($userid);
-                        if (!empty($odburl)) {
-                            $metadata = json_encode(['odburl' => $odburl]);
-                        }
+                // Create o365_object record if it does not exist.
+                $tenant = utils::get_tenant_for_user($userid);
+                $metadata = '';
+                if (!empty($tenant)) {
+                    // Additional tenant - get ODB url.
+                    $odburl = utils::get_odburl_for_user($userid);
+                    if (!empty($odburl)) {
+                        $metadata = json_encode(['odburl' => $odburl]);
                     }
-                    $now = time();
-                    $userobjectdata = (object)[
-                        'type' => 'user',
-                        'subtype' => '',
-                        'objectid' => $o365user->objectid,
-                        'o365name' => str_replace('#ext#', '#EXT#', $o365user->upn),
-                        'moodleid' => $userid,
-                        'tenant' => $tenant,
-                        'metadata' => $metadata,
-                        'timecreated' => $now,
-                        'timemodified' => $now,
-                    ];
-                    $userobjectdata->id = $DB->insert_record('local_o365_objects', $userobjectdata);
                 }
+                $now = time();
+                $userobjectdata = (object)[
+                    'type' => 'user',
+                    'subtype' => '',
+                    'objectid' => $o365user->objectid,
+                    'o365name' => str_replace('#ext#', '#EXT#', $o365user->upn),
+                    'moodleid' => $userid,
+                    'tenant' => $tenant,
+                    'metadata' => $metadata,
+                    'timecreated' => $now,
+                    'timemodified' => $now,
+                ];
+                $userobjectdata->id = $DB->insert_record('local_o365_objects', $userobjectdata);
             }
 
             // Sync profile photo and timezone.
@@ -841,6 +838,8 @@ class observers {
 
         $eventdata = $event->get_data();
 
+        $cachespurgeneeded = false;
+
         if ($eventdata['other']['plugin'] == 'auth_oidc') {
             switch ($eventdata['other']['name']) {
                 case 'clientid':
@@ -872,6 +871,8 @@ class observers {
                     unset_config('apptokens', 'local_o365');
                     unset_config('adminconsent', 'local_o365');
                     unset_config('azuresetupresult', 'local_o365');
+
+                    $cachespurgeneeded = true;
             }
         }
 
@@ -882,15 +883,22 @@ class observers {
 
             // Clear user records in local_o365_objects table.
             $DB->delete_records('local_o365_objects', ['type' => 'user']);
+
+            $cachespurgeneeded = true;
         }
 
-        // If connection method is changed from system API user to applicationa access, system API user token needs to be deleted.
+        // If connection method is changed from system API user to application access, system API user token needs to be deleted.
         if ($eventdata['other']['plugin'] == 'local_o365' && $eventdata['other']['name'] == 'enableapponlyaccess' &&
             $eventdata['other']['oldvalue'] == '0' && $eventdata['other']['value'] == '1') {
             unset_config('systemtokens', 'local_o365');
+
+            $cachespurgeneeded = true;
         }
 
-        purge_all_caches();
+        // Purge caches if needed.
+        if ($cachespurgeneeded) {
+            purge_all_caches();
+        }
 
         return true;
     }
